Fix a bug with classic obj (not dataclasses) as args where repr/str are not implemented.
So signature did NOT reflect changes in the obj.
Now obj are handled because signature is computed based on obj.__dict__.
This is not perfect as:

- Some private attr of obj are not relevant and should not be monitored. They will create lot of cache entries. Default cache max_size=255 should limit the perf impact.  But at least client will get a return value from the fun call and not always the same value from the cache.

- Nested obj signature won't be computed.

One can think obj.__hash__ must be better.
But testing shows default __hash_ (which is rarely override) doesn't change in some situations like the one describe in tests.